+++ app/src/main/java/com/nbscollege/fitnessapp/mainscreen/categorycard/AbsExerciseList.kt
@@ -1,5 +1,6 @@
 package com.nbscollege.fitnessapp.mainscreen.categorycard
 
+//ExerciseList is now = to AbsExerciseList
 import android.annotation.SuppressLint
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -47,7 +48,6 @@ import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.LottieConstants
 import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
-//ExerciseList is now = to AbsExerciseList
 import com.nbscollege.fitnessapp.R
 
 
@@ -107,6 +107,7 @@ fun AbsExerciseList (navController: NavController) {
                 .fillMaxSize()
                 .padding(innerPadding)
         ) {
+
             Column(
                 modifier = Modifier.fillMaxSize(),
                 horizontalAlignment = Alignment.CenterHorizontally,
@@ -129,7 +130,7 @@ fun AbsExerciseList (navController: NavController) {
                         )
                     }
                     item {
-                        com.nbscollege.fitnessapp.mainscreen.dataclass.AbsExerciseList.forEachIndexed() { index, exercise ->
+                        com.nbscollege.fitnessapp.mainscreen.dataclass.ExerciseList.forEachIndexed { index, exercise ->
                             Card(
                                 modifier = Modifier
                                     .fillMaxWidth(),
@@ -144,13 +145,442 @@ fun AbsExerciseList (navController: NavController) {
                                         .height(90.dp)
                                         .clip(RoundedCornerShape(16.dp))
                                 ) {
+
+                                    Button(
+                                        onClick = {
+                                            navController.navigate("CategoryDetails/$index")
+                                        },
+                                        shape = RoundedCornerShape(1.dp),
+                                        modifier = Modifier
+                                            .fillMaxWidth()
+                                            .clip(RoundedCornerShape(16.dp))
+                                            .height(90.dp),
+                                        colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent)
+                                    ) {
+
+                                        Column(
+                                            modifier = Modifier
+                                                .fillMaxSize()
+                                                .padding(1.dp),
+                                            verticalArrangement = Arrangement.Center
+                                        ) {
+                                            Text(
+                                                text = exercise.title,
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                               ,
+                                                textAlign = TextAlign.Center
+                                            )
+
+                                            Text(
+
+                                                "${exercise.time} Seconds",
+                                                        color = Color.Black,
+                                                        fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                             ,
+                                                textAlign = TextAlign.Center,
+
+                                            )
+
+                                        }
+
+
+
+                                    }
+
+
+                                }
+
+                            }
+                            Spacer(modifier = Modifier.height(20.dp))
+                        }
+
+
+                    }
+
+                }
+
+
+
+            }
+        }
+
+    }
+
+}
+
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ArmExerciseList (navController: NavController) {
+
+    val state = rememberScrollState()
+    LaunchedEffect(Unit) { state.animateScrollTo(100) }
+
+    Scaffold(
+        topBar = {
+            Box(
+                modifier = Modifier
+                    .background(Color.White)
+                    .fillMaxWidth()
+            ) {
+                SmallFloatingActionButton(
+                    onClick = {
+                        navController.navigate("HomeScreen")
+                    },
+                    containerColor = Color.Transparent,
+                    modifier = Modifier
+                        .padding(start = 5.dp, end = 5.dp)
+                        .zIndex(3f)
+                ) {
+                    Icon(
+                        Icons.Filled.KeyboardArrowLeft, "Back",
+                        modifier = Modifier.size(40.dp),
+                        tint = Color.White
+                    )
+                }
+                Image(
+                    painter = painterResource(id = R.drawable.arm),
+                    contentDescription = null,
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(150.dp),
+                )
+            }
+        },
+//        bottomBar = {
+//            BottomAppBar {
+//
+//            }
+//        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .background(Color.White)
+                .fillMaxSize()
+                .padding(innerPadding)
+        ) {
+            Column(
+                modifier = Modifier.fillMaxSize(),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Center
+            ) {
+                LazyColumn(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(10.dp),
+                ) {
+                    item {
+                        Spacer(modifier = Modifier.height(20.dp))
+                        Text(
+                            text = "EXERCISES",
+                            fontWeight = FontWeight.Bold,
+                            fontSize = 20.sp,
+                            color = Color.Black,
+                            modifier = Modifier
+                        )
+                    }
+                    item {
+                        com.nbscollege.fitnessapp.mainscreen.dataclass.ArmExerciseList.forEachIndexed { index, exercise ->
+                            Card(
+                                modifier = Modifier
+                                    .fillMaxWidth(),
+                                elevation = CardDefaults.cardElevation(defaultElevation = 10.dp)
+                            ) {
+                                Box(
+                                    modifier = Modifier
+                                        .fillMaxWidth()
+                                        .background(Color.White)
+                                        .height(90.dp)
+                                        .clip(RoundedCornerShape(16.dp))
+                                ) {
+                                    Button(
+                                        onClick = {
+                                            navController.navigate("CategoryDetails/$index")
+                                        },
+                                        shape = RoundedCornerShape(1.dp),
+                                        modifier = Modifier
+                                            .fillMaxWidth()
+                                            .clip(RoundedCornerShape(16.dp))
+                                            .height(90.dp),
+                                        colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent)
+                                    ) {
+                                        Column(
+                                            modifier = Modifier
+                                                .fillMaxSize()
+                                                .padding(1.dp),
+                                            verticalArrangement = Arrangement.Center
+                                        ) {
+                                            Text(
+                                                text = exercise.title,
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                                ,
+                                                textAlign = TextAlign.Center
+                                            )
+                                            Text(
+                                                "${exercise.time} Seconds",
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                                ,
+                                                textAlign = TextAlign.Center,
+                                                )
+                                        }
+                                    }
+
+                                }
+
+                            }
+                            Spacer(modifier = Modifier.height(20.dp))
+                        }
+
+                    }
+
+                }
+
+            }
+        }
+
+    }
+
+}
+
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ChestExerciseList (navController: NavController) {
+
+    val state = rememberScrollState()
+    LaunchedEffect(Unit) { state.animateScrollTo(100) }
+
+    Scaffold(
+        topBar = {
+            Box(
+                modifier = Modifier
+                    .background(Color.White)
+                    .fillMaxWidth()
+            ) {
+                SmallFloatingActionButton(
+                    onClick = {
+                        navController.navigate("HomeScreen")
+                    },
+                    containerColor = Color.Transparent,
+                    modifier = Modifier
+                        .padding(start = 5.dp, end = 5.dp)
+                        .zIndex(3f)
+                ) {
+                    Icon(
+                        Icons.Filled.KeyboardArrowLeft, "Back",
+                        modifier = Modifier.size(40.dp),
+                        tint = Color.White
+                    )
+                }
+                Image(
+                    painter = painterResource(id = R.drawable.arm),
+                    contentDescription = null,
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(150.dp),
+                )
+            }
+        },
+//        bottomBar = {
+//            BottomAppBar {
+//
+//            }
+//        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .background(Color.White)
+                .fillMaxSize()
+                .padding(innerPadding)
+        ) {
+            Column(
+                modifier = Modifier.fillMaxSize(),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Center
+            ) {
+                LazyColumn(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(10.dp),
+                ) {
+                    item {
+                        Spacer(modifier = Modifier.height(20.dp))
+                        Text(
+                            text = "EXERCISES",
+                            fontWeight = FontWeight.Bold,
+                            fontSize = 20.sp,
+                            color = Color.Black,
+                            modifier = Modifier
+                        )
+                    }
+                    item {
+                        com.nbscollege.fitnessapp.mainscreen.dataclass.ChestExerciseList.forEachIndexed { index, exercise ->
+                            Card(
+                                modifier = Modifier
+                                    .fillMaxWidth(),
+                                elevation = CardDefaults.cardElevation(defaultElevation = 10.dp)
+                            ) {
+                                Box(
+                                    modifier = Modifier
+                                        .fillMaxWidth()
+                                        .background(Color.White)
+                                        .height(90.dp)
+                                        .clip(RoundedCornerShape(16.dp))
+                                ) {
+                                    Button(
+                                        onClick = {
+                                            navController.navigate("CategoryDetails/$index")
+                                        },
+                                        shape = RoundedCornerShape(1.dp),
+                                        modifier = Modifier
+                                            .fillMaxWidth()
+                                            .clip(RoundedCornerShape(16.dp))
+                                            .height(90.dp),
+                                        colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent)
+                                    ) {
+                                        Column(
+                                            modifier = Modifier
+                                                .fillMaxSize()
+                                                .padding(1.dp),
+                                            verticalArrangement = Arrangement.Center
+                                        ) {
+                                            Text(
+                                                text = exercise.title,
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                                ,
+                                                textAlign = TextAlign.Center
+                                            )
+                                            Text(
+                                                "${exercise.time} Seconds",
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                                ,
+                                                textAlign = TextAlign.Center,
+                                            )
+                                        }
+                                    }
+
+                                }
+
+                            }
+                            Spacer(modifier = Modifier.height(20.dp))
+                        }
+
+                    }
+
+                }
+
+            }
+        }
+
+    }
+}
+
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LegExerciseList (navController: NavController) {
+
+    val state = rememberScrollState()
+    LaunchedEffect(Unit) { state.animateScrollTo(100) }
+
+    Scaffold(
+        topBar = {
+            Box(
+                modifier = Modifier
+                    .background(Color.White)
+                    .fillMaxWidth()
+            ) {
+                SmallFloatingActionButton(
+                    onClick = {
+                        navController.navigate("HomeScreen")
+                    },
+                    containerColor = Color.Transparent,
+                    modifier = Modifier
+                        .padding(start = 5.dp, end = 5.dp)
+                        .zIndex(3f)
+                ) {
+                    Icon(
+                        Icons.Filled.KeyboardArrowLeft, "Back",
+                        modifier = Modifier.size(40.dp),
+                        tint = Color.White
+                    )
+                }
+                Image(
+                    painter = painterResource(id = R.drawable.leg),
+                    contentDescription = null,
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(150.dp),
+                )
+            }
+        },
+//        bottomBar = {
+//            BottomAppBar {
+//
+//            }
+//        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .background(Color.White)
+                .fillMaxSize()
+                .padding(innerPadding)
+        ) {
+            Column(
+                modifier = Modifier.fillMaxSize(),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Center
+            ) {
+                LazyColumn(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(10.dp),
+                ) {
+                    item {
+                        Spacer(modifier = Modifier.height(20.dp))
+                        Text(
+                            text = "LEG EXERCISES",
+                            fontWeight = FontWeight.Bold,
+                            fontSize = 20.sp,
+                            color = Color.Black,
+                            modifier = Modifier
+                        )
+                    }
+                    item {
+                        com.nbscollege.fitnessapp.mainscreen.dataclass.LegExerciseList.forEachIndexed { index, exercise ->
+                            val composition by rememberLottieComposition(spec = LottieCompositionSpec.RawRes(
+                                com.nbscollege.fitnessapp.mainscreen.dataclass.LegExerciseList[index].animation
+                            ))
+
+                            Card(
+                                modifier = Modifier
+                                    .fillMaxWidth(),
+                                elevation = CardDefaults.cardElevation(defaultElevation = 10.dp)
+                            ) {
+                                Box(
+                                    modifier = Modifier
+                                        .fillMaxWidth()
+                                        .background(Color.White)
+                                        .height(90.dp)
+                                        .clip(RoundedCornerShape(16.dp))
+                                ) {
                                     Row (
                                         modifier = Modifier
                                             .padding(start = 12.dp),
                                         horizontalArrangement = Arrangement.Start
                                     ) {
-                                        val composition by rememberLottieComposition(spec = LottieCompositionSpec.RawRes(
-                                            com.nbscollege.fitnessapp.mainscreen.dataclass.AbsExerciseList[index].animation))
+
 
                                         val preloaderProgress by animateLottieCompositionAsState(
                                             composition,
@@ -161,13 +591,13 @@ fun AbsExerciseList (navController: NavController) {
                                         LottieAnimation(
                                             modifier = Modifier.size(80.dp),
                                             progress = preloaderProgress,
-                                            composition = composition,
+                                            composition =  composition,
 
                                             )
 
                                         Button(
                                             onClick = {
-                                                navController.navigate("AbsDetails/$index")
+                                                navController.navigate("CategoryDetails/$index")
                                             },
                                             shape = RoundedCornerShape(1.dp),
                                             modifier = Modifier
@@ -176,11 +606,10 @@ fun AbsExerciseList (navController: NavController) {
                                                 .height(90.dp),
                                             colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent)
                                         ) {
-
                                             Column(
                                                 modifier = Modifier
                                                     .fillMaxSize()
-                                                    .padding(1.dp),
+                                                    .padding(start = 1.dp),
                                                 verticalArrangement = Arrangement.Center
                                             ) {
                                                 Text(
@@ -190,33 +619,170 @@ fun AbsExerciseList (navController: NavController) {
                                                     modifier = Modifier,
                                                     textAlign = TextAlign.Center
                                                 )
-
                                                 Text(
-
                                                     "${exercise.time} Seconds",
                                                     color = Color.Black,
                                                     fontWeight = FontWeight.Bold,
                                                     modifier = Modifier,
                                                     textAlign = TextAlign.Center,
-
-                                                    )
+                                                )
                                             }
                                         }
+
                                     }
+
+
+
+
                                 }
+
                             }
                             Spacer(modifier = Modifier.height(20.dp))
                         }
+
                     }
+
                 }
+
             }
         }
+
     }
 }
 
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ShoulderExerciseList (navController: NavController) {
 
+    val state = rememberScrollState()
+    LaunchedEffect(Unit) { state.animateScrollTo(100) }
 
+    Scaffold(
+        topBar = {
+            Box(
+                modifier = Modifier
+                    .background(Color.White)
+                    .fillMaxWidth()
+            ) {
+                SmallFloatingActionButton(
+                    onClick = {
+                        navController.navigate("HomeScreen")
+                    },
+                    containerColor = Color.Transparent,
+                    modifier = Modifier
+                        .padding(start = 5.dp, end = 5.dp)
+                        .zIndex(3f)
+                ) {
+                    Icon(
+                        Icons.Filled.KeyboardArrowLeft, "Back",
+                        modifier = Modifier.size(40.dp),
+                        tint = Color.White
+                    )
+                }
+                Image(
+                    painter = painterResource(id = R.drawable.shoulderback),
+                    contentDescription = null,
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(150.dp),
+                )
+            }
+        },
+//        bottomBar = {
+//            BottomAppBar {
+//
+//            }
+//        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .background(Color.White)
+                .fillMaxSize()
+                .padding(innerPadding)
+        ) {
+            Column(
+                modifier = Modifier.fillMaxSize(),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Center
+            ) {
+                LazyColumn(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(10.dp),
+                ) {
+                    item {
+                        Spacer(modifier = Modifier.height(20.dp))
+                        Text(
+                            text = "EXERCISES",
+                            fontWeight = FontWeight.Bold,
+                            fontSize = 20.sp,
+                            color = Color.Black,
+                            modifier = Modifier
+                        )
+                    }
+                    item {
+                        com.nbscollege.fitnessapp.mainscreen.dataclass.ShoulderExerciseList.forEachIndexed { index, exercise ->
+                            Card(
+                                modifier = Modifier
+                                    .fillMaxWidth(),
+                                elevation = CardDefaults.cardElevation(defaultElevation = 10.dp)
+                            ) {
+                                Box(
+                                    modifier = Modifier
+                                        .fillMaxWidth()
+                                        .background(Color.White)
+                                        .height(90.dp)
+                                        .clip(RoundedCornerShape(16.dp))
+                                ) {
+                                    Button(
+                                        onClick = {
+                                            navController.navigate("CategoryDetails/$index")
+                                        },
+                                        shape = RoundedCornerShape(1.dp),
+                                        modifier = Modifier
+                                            .fillMaxWidth()
+                                            .clip(RoundedCornerShape(16.dp))
+                                            .height(90.dp),
+                                        colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent)
+                                    ) {
+                                        Column(
+                                            modifier = Modifier
+                                                .fillMaxSize()
+                                                .padding(1.dp),
+                                            verticalArrangement = Arrangement.Center
+                                        ) {
+                                            Text(
+                                                text = exercise.title,
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                                ,
+                                                textAlign = TextAlign.Center
+                                            )
+                                            Text(
+                                                "${exercise.time} Seconds",
+                                                color = Color.Black,
+                                                fontWeight = FontWeight.Bold,
+                                                modifier = Modifier
+                                                ,
+                                                textAlign = TextAlign.Center,
+                                            )
+                                        }
+                                    }
 
+                                }
 
+                            }
+                            Spacer(modifier = Modifier.height(20.dp))
+                        }
 
+                    }
 
+                }
+
+            }
+        }
+
+    }
+}
